refactor: change return value of generate_log()
generate_log() used to return a structure of form:
  1. {"property": x, "val1": a} - for unchanged, removed and added
    properties.
  2. {"property": x, "val1": a, "val2": b} - for changed properties.

Now generate_log() returns a structure of form:
  {"property": x, values:{val1: a, ...}}
Where all values (2 for changed property and 1 for other) are stored in
a separate dict inside the outer dict.

Functions that process the output of generate_log() in any way were
also changed accordingly.